vfio/pci: Intel IGD graphics support
In order to support device assignment of Intel IGD graphics, we need
support three quirks, which are enabled through vfio using device
specific regions.

The first quirk is to expose the OpRegion of the host.  This contains
a Video BIOS Table (VBT), that often provides graphics modes that are
useful for laptops.  We store the OpRegion data from vfio in a buffer
and pass it to SeaBIOS via fw_cfg.  The ASL Storage register (0xFC) is
made writable so the BIOS can use this register to inform the guest OS
of the allocated OpRegion in VM memory.  If we were to find a use for
any passthrough of the host OpRegion into the guest (ie. volatile
status bits and whatnot), we could use the ASLS register write as a
trigger to map the host OpRegion into the VM.

The second quirk is to copy some of the revisions and identification
fields from PCI config space on the host bridge into the guest.  Vfio
provides a separate region for providing read-only access to the host
bridge config space to make this possible.

The third and final quirk is to create an LPC/ISA bridge in the VM at
address 00:1f.0 and also populate some of its config space with host
data for revision and identification.  Vfio provides yet another
region for read-only access to this configuration space on the host.

For Broadwell and newer graphics and new enough guest graphics
drivers (at least in Windows), these latter two quirks supposedly
aren't necessary, but since we don't know what driver is running in
the guest, we currently enable these regardless of the graphics
revision.

Signed-off-by: Alex Williamson <alex.williamson@redhat.com>