vfio/pci: Intel IGD stolen memory quirk
The IGD vBIOS really wants to use stolen memory, which it programs
into the device using I/O port BAR4.  This is a typical dword index
and data register where indexes below 0x400 seem to be writing stolen
memory addresses.  The vBIOS apparently comes up with the address of
stolen memory itself or it's written into the ROM, because it's the
host stolen memory location.  Since we really don't want to identity
map the host stolen memory into the guest, we intercept these writes
and instead program the device with memory in the VM, reserved for us
by SeaBIOS.  The result is that the vBIOS works as intended and we
don't get any DMAR faults or VM memory corruption by the graphics
device trying to use host memory addresses.  Of course being able to
do this implies that the stolen memory is fire-and-forget for the
vBIOS, ie. it doesn't actually access the memory range itself.  This
does seem to be the case since it works.

Signed-off-by: Alex Williamson <alex.williamson@redhat.com>